Relocate `Collection` objects
More use of `with_x(FnOnce)` pattern
