+++ src/core/activity.rs
@@ -40,6 +40,7 @@ pub struct Activity {
 
 impl ActivityBuilder {
 
+    // TODO: macro
     pub fn with_base<F>(&mut self, build_fn: F) -> &mut Self
         where F: FnOnce(&mut ObjectBuilder) -> &mut ObjectBuilder
     {
@@ -61,6 +62,12 @@ impl ActivityBuilder {
         self.actor(Some(build_fn(&mut base_builder).build().unwrap()))
     }
 
+    pub fn with_target<F>(&mut self, build_fn: F) -> &mut Self
+        where F: FnOnce(&mut ObjectBuilder) -> &mut ObjectBuilder
+    {
+        let mut base_builder = ObjectBuilder::default();
+        self.target(Some(build_fn(&mut base_builder).build().unwrap()))
+    }
     /// Instances of [IntransitiveActivity] are a subtype of [Activity] representing
     /// intransitive actions. The object property is therefore inappropriate for
     /// these activities.
